organise singletons into `singletons/`
