+++ src/singletons/alchemyWeb3Instance.ts
@@ -1,6 +1,6 @@
 import {createAlchemyWeb3} from '@alch/alchemy-web3';
 
-import {ALCHEMY_WS_URL} from './config';
+import {ALCHEMY_WS_URL} from '../config';
 
 /**
  * Web3 instance singleton
