+++ src/singletons/index.ts
@@ -1,0 +1,2 @@
+export * from './alchemyWeb3Instance';
+export * from './prismaClientInstance';
