+++ src/singletons/prismaClientInstance.ts
@@ -7,6 +7,4 @@
 
 import {PrismaClient} from '@prisma/client';
 
-const prisma = new PrismaClient();
-
-export default prisma;
+export const prisma = new PrismaClient();
